try using FloatField instead

+++ content/models.py
@@ -2,7 +2,6 @@ from django.db import models
 from django.forms import ModelForm
 from django.contrib.auth.models import User
 from taggit.managers import TaggableManager
-from decimal import Decimal
 
 class Dict(models.Model): #collection of TagVals or Voters (associated by ForeignKeys)
     name = models.CharField(max_length=500)
@@ -14,7 +13,7 @@ class TagVal(models.Model): #holds a tag and associated value, like posts, doubl
     container = models.ForeignKey(Dict, db_index=True)
     tag = models.CharField(max_length=100, db_index=True)
     val = models.IntegerField(default=0, db_index=True)
-    dval = models.DecimalField(max_digits=8, decimal_places=3, null=True, db_index=True)
+    dval = models.FloatField(null=True, db_index=True)
 
     def __unicode__(self):
         return self.tag+', '+str(self.val)
